refactor: moved and renamed test classes around the member_credit_settle view.
